Move files, add error handling
